test matrices moved to testg abstract and pass
